+++ R/chunk_processing.R
@@ -1,27 +1,60 @@
 ## Chunk processing
+#' Which chunks are safe to exclude
+#'
+#' Check to see which chunks do not have include = FALSE
+#'
+#' @param chunk_options Chunk options vector
+#'
+#' @return Logical vector of chunks that can be safely unnamed
+#' @details We want to make sure chunks invovled in flipbookr remain named,
+#' so we don't re-name them and then break the label that allows flipbooks to be constructed.
+#'
+check_chunk_include <- function(chunk_options = chunk_options) {
+  x <- stringr::str_remove_all(chunk_options, " ")
+  ## Chunks not safe to unname
+  stringr::str_detect(x, "include=FALSE")
+}
 
+#' Unname all unnameable chunks
+#'
+#' Unname chunks except 'setup' and flipbookr chunks that have include=FALSE
+#'
+#' @param path Path to Rmd file
+#' @param chunk_name_prefix Character string with prefix of chunknames that will be removed.
+#' Default: NULL (indicating all chunknames will be removed except the one named 'setup' and
+#' any chunks with the include = FALSE option, to protect flipbookr references.)
+#'
+#' @return Unnames chunks, except setup and include=FALSE chunks
+#' @export
+#'
+kjh_unname_chunks <- function (path, chunk_name_prefix = NULL) {
 
-# kjh_unname_chunk <- function (path, chunk_name_prefix = NULL)
-# {
-#   lines <- readLines(path)
-#   chunk_headers_info <- get_chunk_info(lines)
-#   if (is.null(chunk_headers_info)) {
-#     return(invisible("TRUE"))
-#   }
-#   if (is.null(chunk_name_prefix)) {
-#     chunk_headers_info$name[chunk_headers_info$name != "setup"] <- ""
-#   }
-#   else {
-#     del_labels <- strtrim(chunk_headers_info$name, nchar(chunk_name_prefix)) %in%
-#       chunk_name_prefix
-#     setup_label <- !(chunk_headers_info$name %in% "setup")
-#     del_labels <- del_labels & setup_label
-#     chunk_headers_info$name[del_labels] <- ""
-#   }
-#   newlines <- re_write_headers(chunk_headers_info)
-#   lines[newlines$index] <- newlines$line
-#   writeLines(lines, path)
-# }
+  lines <- readLines(path)
+  chunk_headers_info <- namer:::get_chunk_info(lines)
+
+  chunk_options <- chunk_headers_info$options
+  check_include <- check_chunk_include(chunk_options)
+
+  if (is.null(chunk_headers_info)) {
+    return(invisible("TRUE"))
+  }
+  if (is.null(chunk_name_prefix)) {
+    ind <- !((chunk_headers_info$name %in% "setup") | check_include)
+    chunk_headers_info$name[ind] <- ""
+  }
+  else {
+    del_labels <- strtrim(chunk_headers_info$name,
+                          nchar(chunk_name_prefix)) %in% chunk_name_prefix
+
+    setup_label <- !(chunk_headers_info$name %in% c("setup") |
+                     check_include)
+    del_labels <- del_labels & setup_label
+    chunk_headers_info$name[del_labels] <- ""
+  }
+  newlines <- namer:::re_write_headers(chunk_headers_info)
+  lines[newlines$index] <- newlines$line
+  writeLines(lines, path)
+}
 
 #' Unname all chunks
 #'
@@ -29,8 +62,9 @@
 #'
 #' @param indir The input directory, default "slides"
 #'
-#' @return All the Rmd files with their chunk names removed
-#' @details Recurses 1 level (i.e. subdirs) by default. Right now this will break any references made with chunk_reveal!
+#' @details All the Rmd files with their chunk names removed, except chunks labeled
+#' 'setup' and chunks where include=FALSE, to protect chunk_reveal calls.
+#' Recurses 1 level (i.e. subdirs) by default.
 #' @export
 #' @examples
 #' \dontrun{
@@ -49,7 +83,7 @@ kjh_unname_all_chunks <- function(indir = "slides") {
 
   fnames <- as.vector(dplyr::pull(fnames))
 
-  purrr:::walk(fnames, namer::unname_chunks)
+  purrr:::walk(fnames, kjh_unname_chunks)
 }