+++ NEWS.md
@@ -1,3 +1,11 @@
+# kjhslides 0.0.3.9000
+
+Some helper functions for a Myriad theme.
+
+# kjhslides 0.0.2.9000
+
+* Protect `chunk_reveal()` references when batch unnaming/renaming chunks.
+
 # kjhslides 0.0.1.9000
 
 * Added a `NEWS.md` file to track changes to the package.